move nextschedule to consumables
